refactor(cli): extraer ingesta elástica y parsing de args del orchestrator (#449)
Mueve la ingesta vectorial elástica (run_elastic_ingestion,
build_elastic_ingestion) a cli/elastic.rs, y handle_completions y
parse_asset_naming a cli/parse.rs. orchestrator re-importa los items movidos
para que call sites y tests unitarios resuelvan sin cambios; handle_completions
se re-exporta para preservar orchestrator::handle_completions.

parse_asset_h2_profile se mantiene en orchestrator: emite un tracing::warn!
cuyo target (modulo) y archivo de origen estan fijados por el snapshot
behavioral unknown_h2_profile_stderr; moverlo cambiaria esa salida observable.
