+++ crates/webfang_core/src/cli/mod.rs
@@ -7,9 +7,11 @@ pub mod args;
 pub mod commands;
 pub mod completions;
 pub mod config;
+pub mod elastic;
 pub mod error;
 pub mod export_flow;
 pub mod orchestrator;
+pub mod parse;
 pub mod preflight;
 pub mod scrape_flow;
 pub mod summary;
